Creates new class to store only hash data for communications.
-SenderData class encapsulates the master index, block size, and hashes. It also avoids sending the entire byte array.
-Changes BlockSize to be of type int for convenience.

+++ src/ec504project/application/BlockMatcher.java
@@ -13,10 +13,10 @@ public class BlockMatcher {
 	private HashMap<Integer, HashObject> receiverHashes;
 	public ArrayList<ReconcileStep> receiverSteps;
 	
-	private long blockSize;
+	private int blockSize;
 	
 	
-	public BlockMatcher(File input, HashMap<Integer, HashObject> receiverHashes, long blockSize) {
+	public BlockMatcher(File input, HashMap<Integer, HashObject> receiverHashes, int blockSize) {
 		this.receiverHashes = receiverHashes;
 		this.blockSize = blockSize;
